+++ demo/src/main/java/raven/modal/demo/Demo.java
@@ -3,6 +3,7 @@ package raven.modal.demo;
 import com.formdev.flatlaf.FlatClientProperties;
 import com.formdev.flatlaf.FlatLaf;
 import com.formdev.flatlaf.fonts.roboto.FlatRobotoFont;
+import com.formdev.flatlaf.util.FontUtils;
 import raven.modal.Drawer;
 import raven.modal.demo.menu.MyDrawerBuilder;
 import raven.modal.demo.system.FormManager;
@@ -13,7 +14,7 @@ import java.awt.*;
 
 public class Demo extends JFrame {
 
-    public static final String DEMO_VERSION = "2.4.1-SNAPSHOT";
+    public static final String DEMO_VERSION = "2.5.0";
 
     public Demo() {
         init();
@@ -22,7 +23,7 @@ public class Demo extends JFrame {
     private void init() {
         setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
         getRootPane().putClientProperty(FlatClientProperties.FULL_WINDOW_CONTENT, true);
-        Drawer.installDrawer(this, new MyDrawerBuilder());
+        Drawer.installDrawer(this, MyDrawerBuilder.getInstance());
         FormManager.install(this);
         setSize(new Dimension(1366, 768));
         setLocationRelativeTo(null);
@@ -32,7 +33,7 @@ public class Demo extends JFrame {
         DemoPreferences.init();
         FlatRobotoFont.install();
         FlatLaf.registerCustomDefaultsSource("raven.modal.demo.themes");
-        UIManager.put("defaultFont", new Font(FlatRobotoFont.FAMILY, Font.PLAIN, 13));
+        UIManager.put("defaultFont", FontUtils.getCompositeFont(FlatRobotoFont.FAMILY, Font.PLAIN, 13));
         DemoPreferences.setupLaf();
         EventQueue.invokeLater(() -> new Demo().setVisible(true));
     }