+++ modal-dialog/src/main/java/raven/modal/drawer/simple/footer/SimpleFooterStyle.java
@@ -12,4 +12,8 @@ public abstract class SimpleFooterStyle {
 
     public void styleComponent(JComponent component, int styleType) {
     }
+
+    public int getTextGap() {
+        return 3;
+    }
 }