+++ modal-dialog/src/main/java/raven/modal/utils/FlatLafStyleUtils.java
@@ -1,8 +1,11 @@
 package raven.modal.utils;
 
 import com.formdev.flatlaf.FlatClientProperties;
+import com.formdev.flatlaf.ui.FlatStylingSupport;
+import com.formdev.flatlaf.ui.FlatUIUtils;
 
 import javax.swing.*;
+import java.awt.*;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.StringJoiner;
@@ -43,6 +46,17 @@ public class FlatLafStyleUtils {
         return mapToString(newStyleMap);
     }
 
+    public static <T> T getStyleValue(JComponent com, String key, Class<T> type) {
+        Object style = FlatStylingSupport.getStyle(com);
+        if (style != null) {
+            Object value = FlatStylingSupport.parse(style.toString()).get(key);
+            if (type.isInstance(value)) {
+                return (T) value;
+            }
+        }
+        return null;
+    }
+
     public static Map<String, String> styleToMap(String style) {
         Map<String, String> mapStyle = new HashMap<>();
         if (style != null) {
@@ -63,5 +77,13 @@ public class FlatLafStyleUtils {
             joiner.add(entry.getKey() + ":" + entry.getValue());
         }
         return joiner.toString();
+    }
+
+    public static String appendMargin(JComponent button, Insets margin) {
+        Insets defaultInset = FlatLafStyleUtils.getStyleValue(button, "margin", Insets.class);
+        if (defaultInset != null) {
+            margin = FlatUIUtils.addInsets(margin, defaultInset);
+        }
+        return margin.top + "," + margin.left + "," + margin.bottom + "," + margin.right;
     }
 }